Move MutNullableDom to script_bindings.
